shading choosing function
the function that chooses the shading matrix based on the module's
position is written and seems to function.

Still need to figure out how to pass row and column variables to it
during every module instantiation

Next up: move the shading function from envelope scope into module
scope. That's coupled surgery, though, as opposed to this, which was
stand-alone. let's wrap our branches together before we do that.