everywhere and nowhere
Tried to get a selector working for the shading LUT. but can't get it
running without it using algorithm functions, and they don't play nice
with whatever is running the LUT selector component (named
ShadingLUT_Twelve. I did figure out a lot of syntax to brute-force the
selector, tho, and the combitables are pulling in data from a file. Note
- the current method does add A BUNCH of equations to Module...must be a
better way.