Adding temperature outputs at the module flowport
Flowports don't report temperature.  It can be calculate from the
H_flow, where H_flow = m_flow * cp. In the skeleton the m.dot is
calculated from rho*v_flow_in, and at the module it is calculated as I
later learn directly from the m_flow, which can be access as a parameter
from the flowport.