Working on Truing Model
eta_optical is good.  The convection between the receiver and amb had a
delta t of >50. Using the cca temps on module 2&3. the conv, and cond
loss at the receiver are now in ~range.  Balancing has mainly been
looking at the 3 temps on s3m3 and s3m2.  Still need to true heat loss
between runs