added feb20 dataset and updated row#s
the number of rows claimed (in the parents in line 2) wasn't necessarily
accurate in the 2-20, 2-28, 3-19, and 3-23 data, so I checked and fixed
them. Hopefully my fixes cause good and not bad.

ATTENTION: also took out some extra CR's at the end of some of the data
files. put them back if they cause trouble?