Location, Tilt, and Tempt study sync
Have no synced in a long time. This sync includes the simulations run
for nick's defense and the FacadeGeneration data in ...\IC Solar in
Studio\4-ENERGY SIMULATIONS\ModelicaYearLong_OrientationAndTiltStudy. It
also included the temperature and exergy studies we're doing right now.
Those scripts can be found in the Archive/Modelica Scirpting/. Ask JS
for the data from the sims.